Change API base name to connector
