+++ ibis-server/tests/routers/v2/connector/test_bigquery.py
@@ -3,61 +3,21 @@ import base64
 import orjson
 import pytest
 from fastapi.testclient import TestClient
-from sqlalchemy import text
-from testcontainers.mysql import MySqlContainer
 
 from app.main import app
 
 client = TestClient(app)
 
 
-@pytest.fixture(scope="class")
-def mysql(request) -> MySqlContainer:
-    def file_path(path: str) -> str:
-        import os
-
-        return os.path.join(os.path.dirname(__file__), path)
-
-    import pandas as pd
-    import sqlalchemy
-
-    mysql = MySqlContainer("mysql:8.0")
-    mysql.start()
-    connection_url = mysql.get_connection_url()
-    engine = sqlalchemy.create_engine(connection_url)
-    pd.read_parquet(file_path("../../resource/tpch/data/orders.parquet")).to_sql(
-        "orders", engine, index=False
-    )
-    pd.read_parquet(file_path("../../resource/tpch/data/customer.parquet")).to_sql(
-        "customer", engine, index=False
-    )
-    conn = engine.connect()
-    conn.execute(text("ALTER TABLE orders ADD PRIMARY KEY(o_orderkey);"))
-    conn.execute(text("ALTER TABLE customer Add PRIMARY KEY(c_custkey);"))
-    conn.execute(
-        text(
-            "ALTER TABLE orders ADD FOREIGN KEY (o_custkey) REFERENCES customer(c_custkey);"
-        )
-    )
-    conn.close()
-
-    def stop_pg():
-        mysql.stop()
-
-    request.addfinalizer(stop_pg)
-
-    return mysql
-
-
-@pytest.mark.mysql
-class TestMySql:
+@pytest.mark.bigquery
+class TestBigquery:
     manifest = {
         "catalog": "my_catalog",
         "schema": "my_schema",
         "models": [
             {
                 "name": "Orders",
-                "refSql": "select * from orders",
+                "refSql": "select * from tpch_tiny.orders",
                 "columns": [
                     {"name": "orderkey", "expression": "o_orderkey", "type": "integer"},
                     {"name": "custkey", "expression": "o_custkey", "type": "integer"},
@@ -92,7 +52,7 @@ class TestMySql:
             },
             {
                 "name": "Customer",
-                "refSql": "select * from customer",
+                "refSql": "select * from tpch_tiny.customer",
                 "columns": [
                     {"name": "custkey", "expression": "c_custkey", "type": "integer"},
                     {"name": "name", "expression": "c_name", "type": "varchar"},
@@ -105,28 +65,23 @@ class TestMySql:
     manifest_str = base64.b64encode(orjson.dumps(manifest)).decode("utf-8")
 
     @staticmethod
-    def to_connection_info(mysql: MySqlContainer):
+    def get_connection_info():
+        import os
+
         return {
-            "host": mysql.get_container_host_ip(),
-            "port": mysql.get_exposed_port(mysql.port),
-            "user": mysql.username,
-            "password": mysql.password,
-            "database": mysql.dbname,
+            "project_id": os.getenv("TEST_BIG_QUERY_PROJECT_ID"),
+            "dataset_id": "tpch_tiny",
+            "credentials": os.getenv("TEST_BIG_QUERY_CREDENTIALS_BASE64_JSON"),
         }
 
-    @staticmethod
-    def to_connection_url(mysql: MySqlContainer):
-        info = TestMySql.to_connection_info(mysql)
-        return f"mysql://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
-
-    def test_query(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_query(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" LIMIT 1',
+                "sql": 'SELECT * FROM "Orders" ORDER BY orderkey LIMIT 1',
             },
         )
         assert response.status_code == 200
@@ -137,48 +92,31 @@ class TestMySql:
             1,
             370,
             "O",
-            "172799.49",
+            172799.49,
             820540800000,
             "1_370",
             1704153599000,
             1704153599000,
         ]
         assert result["dtypes"] == {
-            "orderkey": "int32",
-            "custkey": "int32",
+            "orderkey": "int64",
+            "custkey": "int64",
             "orderstatus": "object",
-            "totalprice": "object",
+            "totalprice": "float64",
             "orderdate": "object",
             "order_cust_key": "object",
             "timestamp": "datetime64[ns]",
-            "timestamptz": "datetime64[ns]",
+            "timestamptz": "datetime64[ns, UTC]",
         }
 
-    def test_query_with_connection_url(self, mysql: MySqlContainer):
-        connection_url = self.to_connection_url(mysql)
+    def test_query_with_column_dtypes(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
-            json={
-                "connectionInfo": {"connectionUrl": connection_url},
-                "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" LIMIT 1',
-            },
-        )
-        assert response.status_code == 200
-        result = response.json()
-        assert len(result["columns"]) == len(self.manifest["models"][0]["columns"])
-        assert len(result["data"]) == 1
-        assert result["data"][0][0] == 1
-        assert result["dtypes"] is not None
-
-    def test_query_with_column_dtypes(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
-        response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" LIMIT 1',
+                "sql": 'SELECT * FROM "Orders" ORDER BY orderkey LIMIT 1',
                 "columnDtypes": {
                     "totalprice": "float",
                     "orderdate": "datetime64",
@@ -199,11 +137,11 @@ class TestMySql:
             "1996-01-02 00:00:00.000000",
             "1_370",
             "2024-01-01 23:59:59.000000",
-            "2024-01-01 23:59:59.000000",
+            "2024-01-01 23:59:59.000000 UTC",
         ]
         assert result["dtypes"] == {
-            "orderkey": "int32",
-            "custkey": "int32",
+            "orderkey": "int64",
+            "custkey": "int64",
             "orderstatus": "object",
             "totalprice": "float64",
             "orderdate": "object",
@@ -212,10 +150,10 @@ class TestMySql:
             "timestamptz": "object",
         }
 
-    def test_query_without_manifest(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_query_without_manifest(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             json={
                 "connectionInfo": connection_info,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -228,10 +166,10 @@ class TestMySql:
         assert result["detail"][0]["loc"] == ["body", "manifestStr"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_without_sql(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_query_without_sql(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -243,7 +181,7 @@ class TestMySql:
 
     def test_query_without_connection_info(self):
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             json={
                 "manifestStr": self.manifest_str,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -256,10 +194,10 @@ class TestMySql:
         assert result["detail"][0]["loc"] == ["body", "connectionInfo"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_with_dry_run(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_query_with_dry_run(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -269,10 +207,10 @@ class TestMySql:
         )
         assert response.status_code == 204
 
-    def test_query_with_dry_run_and_invalid_sql(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_query_with_dry_run_and_invalid_sql(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/query",
+            url="/v2/connector/bigquery/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -283,10 +221,10 @@ class TestMySql:
         assert response.status_code == 422
         assert response.text is not None
 
-    def test_validate_with_unknown_rule(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_validate_with_unknown_rule(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/validate/unknown_rule",
+            url="/v2/connector/bigquery/validate/unknown_rule",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -299,10 +237,10 @@ class TestMySql:
             == "The rule `unknown_rule` is not in the rules, rules: ['column_is_valid']"
         )
 
-    def test_validate_rule_column_is_valid(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_validate_rule_column_is_valid(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -311,12 +249,10 @@ class TestMySql:
         )
         assert response.status_code == 204
 
-    def test_validate_rule_column_is_valid_with_invalid_parameters(
-        self, mysql: MySqlContainer
-    ):
-        connection_info = self.to_connection_info(mysql)
+    def test_validate_rule_column_is_valid_with_invalid_parameters(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -326,7 +262,7 @@ class TestMySql:
         assert response.status_code == 422
 
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -335,12 +271,10 @@ class TestMySql:
         )
         assert response.status_code == 422
 
-    def test_validate_rule_column_is_valid_without_parameters(
-        self, mysql: MySqlContainer
-    ):
-        connection_info = self.to_connection_info(mysql)
+    def test_validate_rule_column_is_valid_without_parameters(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -350,12 +284,10 @@ class TestMySql:
         assert result["detail"][0]["loc"] == ["body", "parameters"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_validate_rule_column_is_valid_without_one_parameter(
-        self, mysql: MySqlContainer
-    ):
-        connection_info = self.to_connection_info(mysql)
+    def test_validate_rule_column_is_valid_without_one_parameter(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -366,7 +298,7 @@ class TestMySql:
         assert response.text == "Missing required parameter: `columnName`"
 
         response = client.post(
-            url="/v2/ibis/mysql/validate/column_is_valid",
+            url="/v2/connector/bigquery/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -376,33 +308,18 @@ class TestMySql:
         assert response.status_code == 422
         assert response.text == "Missing required parameter: `modelName`"
 
-    def test_metadata_list_tables(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_metadata_list_tables(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/metadata/tables",
+            url="/v2/connector/bigquery/metadata/tables",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
 
-        result = response.json()[0]
-        assert result["name"] is not None
-        assert result["columns"] is not None
-        assert result["primaryKey"] is not None
-        assert result["description"] is not None
-        assert result["properties"] is not None
-
-    def test_metadata_list_constraints(self, mysql: MySqlContainer):
-        connection_info = self.to_connection_info(mysql)
+    def test_metadata_list_constraints(self):
+        connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/mysql/metadata/constraints",
+            url="/v2/connector/bigquery/metadata/constraints",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
-
-        result = response.json()[0]
-        assert result["constraintName"] is not None
-        assert result["constraintType"] is not None
-        assert result["constraintTable"] is not None
-        assert result["constraintColumn"] is not None
-        assert result["constraintedTable"] is not None
-        assert result["constraintedColumn"] is not None
