+++ ibis-server/tests/routers/v2/connector/test_mssql.py
@@ -3,7 +3,7 @@ import base64
 import orjson
 import pytest
 from fastapi.testclient import TestClient
-from testcontainers.postgres import PostgresContainer
+from testcontainers.mssql import SqlServerContainer
 
 from app.main import app
 
@@ -11,7 +11,7 @@ client = TestClient(app)
 
 
 @pytest.fixture(scope="class")
-def postgres(request) -> PostgresContainer:
+def mssql(request) -> SqlServerContainer:
     def file_path(path: str) -> str:
         import os
 
@@ -20,34 +20,36 @@ def postgres(request) -> PostgresContainer:
     import pandas as pd
     import sqlalchemy
 
-    pg = PostgresContainer("postgres:16-alpine")
-    pg.start()
-    psql_url = pg.get_connection_url()
-    engine = sqlalchemy.create_engine(psql_url)
-    pd.read_parquet(file_path("../../resource/tpch/data/orders.parquet")).to_sql(
+    mssql = SqlServerContainer(
+        "mcr.microsoft.com/mssql/server:2019-latest", dialect="mssql+pyodbc"
+    )
+    mssql.start()
+    connection_url = mssql.get_connection_url()
+    engine = sqlalchemy.create_engine(f"{connection_url}?driver=FreeTDS")
+    pd.read_parquet(file_path("../../../resource/tpch/data/orders.parquet")).to_sql(
         "orders", engine, index=False
     )
-    pd.read_parquet(file_path("../../resource/tpch/data/customer.parquet")).to_sql(
+    pd.read_parquet(file_path("../../../resource/tpch/data/customer.parquet")).to_sql(
         "customer", engine, index=False
     )
 
     def stop_pg():
-        pg.stop()
+        mssql.stop()
 
     request.addfinalizer(stop_pg)
 
-    return pg
+    return mssql
 
 
-@pytest.mark.postgres
-class TestPostgres:
+@pytest.mark.mssql
+class TestMSSql:
     manifest = {
         "catalog": "my_catalog",
         "schema": "my_schema",
         "models": [
             {
                 "name": "Orders",
-                "refSql": "select * from public.orders",
+                "refSql": "select * from dbo.orders",
                 "columns": [
                     {"name": "orderkey", "expression": "o_orderkey", "type": "integer"},
                     {"name": "custkey", "expression": "o_custkey", "type": "integer"},
@@ -82,7 +84,7 @@ class TestPostgres:
             },
             {
                 "name": "Customer",
-                "refSql": "select * from public.customer",
+                "refSql": "select * from dbo.customer",
                 "columns": [
                     {"name": "custkey", "expression": "c_custkey", "type": "integer"},
                     {"name": "name", "expression": "c_name", "type": "varchar"},
@@ -95,24 +97,24 @@ class TestPostgres:
     manifest_str = base64.b64encode(orjson.dumps(manifest)).decode("utf-8")
 
     @staticmethod
-    def to_connection_info(pg: PostgresContainer):
+    def to_connection_info(mssql: SqlServerContainer):
         return {
-            "host": pg.get_container_host_ip(),
-            "port": pg.get_exposed_port(pg.port),
-            "user": pg.username,
-            "password": pg.password,
-            "database": pg.dbname,
+            "host": mssql.get_container_host_ip(),
+            "port": mssql.get_exposed_port(mssql.port),
+            "user": mssql.username,
+            "password": mssql.password,
+            "database": mssql.dbname,
         }
 
     @staticmethod
-    def to_connection_url(pg: PostgresContainer):
-        info = TestPostgres.to_connection_info(pg)
-        return f"postgres://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
+    def to_connection_url(mssql: SqlServerContainer):
+        info = TestMSSql.to_connection_info(mssql)
+        return f"mssql://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
 
-    def test_query(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -144,10 +146,13 @@ class TestPostgres:
             "timestamptz": "datetime64[ns, UTC]",
         }
 
-    def test_query_with_connection_url(self, postgres: PostgresContainer):
-        connection_url = self.to_connection_url(postgres)
+    @pytest.mark.skip(
+        reason="ibis does not support mssql using connection url, wait fix PR in ibis repo"
+    )
+    def test_query_with_connection_url(self, mssql: SqlServerContainer):
+        connection_url = self.to_connection_url(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={
                 "connectionInfo": {"connectionUrl": connection_url},
                 "manifestStr": self.manifest_str,
@@ -161,10 +166,10 @@ class TestPostgres:
         assert result["data"][0][0] == 1
         assert result["dtypes"] is not None
 
-    def test_query_with_column_dtypes(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query_with_column_dtypes(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -202,38 +207,10 @@ class TestPostgres:
             "timestamptz": "object",
         }
 
-    def test_query_with_limit(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query_without_manifest(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
-            params={"limit": 1},
-            json={
-                "connectionInfo": connection_info,
-                "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders"',
-            },
-        )
-        assert response.status_code == 200
-        result = response.json()
-        assert len(result["data"]) == 1
-
-        response = client.post(
-            url="/v2/ibis/postgres/query",
-            params={"limit": 1},
-            json={
-                "connectionInfo": connection_info,
-                "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" LIMIT 10',
-            },
-        )
-        assert response.status_code == 200
-        result = response.json()
-        assert len(result["data"]) == 1
-
-    def test_query_without_manifest(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
-        response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={
                 "connectionInfo": connection_info,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -246,10 +223,10 @@ class TestPostgres:
         assert result["detail"][0]["loc"] == ["body", "manifestStr"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_without_sql(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query_without_sql(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -261,7 +238,7 @@ class TestPostgres:
 
     def test_query_without_connection_info(self):
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             json={
                 "manifestStr": self.manifest_str,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -274,10 +251,10 @@ class TestPostgres:
         assert result["detail"][0]["loc"] == ["body", "connectionInfo"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_with_dry_run(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query_with_dry_run(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -287,10 +264,10 @@ class TestPostgres:
         )
         assert response.status_code == 204
 
-    def test_query_with_dry_run_and_invalid_sql(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_query_with_dry_run_and_invalid_sql(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/query",
+            url="/v2/connector/mssql/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -301,10 +278,10 @@ class TestPostgres:
         assert response.status_code == 422
         assert response.text is not None
 
-    def test_validate_with_unknown_rule(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_validate_with_unknown_rule(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/validate/unknown_rule",
+            url="/v2/connector/mssql/validate/unknown_rule",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -317,10 +294,10 @@ class TestPostgres:
             == "The rule `unknown_rule` is not in the rules, rules: ['column_is_valid']"
         )
 
-    def test_validate_rule_column_is_valid(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_validate_rule_column_is_valid(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -330,11 +307,11 @@ class TestPostgres:
         assert response.status_code == 204
 
     def test_validate_rule_column_is_valid_with_invalid_parameters(
-        self, postgres: PostgresContainer
+        self, mssql: SqlServerContainer
     ):
-        connection_info = self.to_connection_info(postgres)
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -344,7 +321,7 @@ class TestPostgres:
         assert response.status_code == 422
 
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -354,11 +331,11 @@ class TestPostgres:
         assert response.status_code == 422
 
     def test_validate_rule_column_is_valid_without_parameters(
-        self, postgres: PostgresContainer
+        self, mssql: SqlServerContainer
     ):
-        connection_info = self.to_connection_info(postgres)
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -369,11 +346,11 @@ class TestPostgres:
         assert result["detail"][0]["msg"] == "Field required"
 
     def test_validate_rule_column_is_valid_without_one_parameter(
-        self, postgres: PostgresContainer
+        self, mssql: SqlServerContainer
     ):
-        connection_info = self.to_connection_info(postgres)
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -384,7 +361,7 @@ class TestPostgres:
         assert response.text == "Missing required parameter: `columnName`"
 
         response = client.post(
-            url="/v2/ibis/postgres/validate/column_is_valid",
+            url="/v2/connector/mssql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -394,32 +371,25 @@ class TestPostgres:
         assert response.status_code == 422
         assert response.text == "Missing required parameter: `modelName`"
 
-    def test_metadata_list_tables(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+    def test_metadata_list_tables(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/metadata/tables",
+            url="/v2/connector/mssql/metadata/tables",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
 
-    def test_metadata_list_constraints(self, postgres: PostgresContainer):
-        connection_info = self.to_connection_info(postgres)
+        result = response.json()[0]
+        assert result["name"] is not None
+        assert result["columns"] is not None
+        assert result["primaryKey"] is not None
+        assert result["description"] is not None
+        assert result["properties"] is not None
+
+    def test_metadata_list_constraints(self, mssql: SqlServerContainer):
+        connection_info = self.to_connection_info(mssql)
         response = client.post(
-            url="/v2/ibis/postgres/metadata/constraints",
+            url="/v2/connector/mssql/metadata/constraints",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
-
-    def test_dry_plan(self):
-        response = client.post(
-            url="/v2/ibis/postgres/dry-plan",
-            json={
-                "manifestStr": self.manifest_str,
-                "sql": 'SELECT orderkey, order_cust_key FROM "Orders" LIMIT 1',
-            },
-        )
-        assert response.status_code == 200
-        assert (
-            response.text
-            == '''"WITH \\"Orders\\" AS (SELECT \\"Orders\\".\\"orderkey\\" AS \\"orderkey\\", \\"Orders\\".\\"custkey\\" AS \\"custkey\\", \\"Orders\\".\\"orderstatus\\" AS \\"orderstatus\\", \\"Orders\\".\\"totalprice\\" AS \\"totalprice\\", \\"Orders\\".\\"orderdate\\" AS \\"orderdate\\", \\"Orders\\".\\"order_cust_key\\" AS \\"order_cust_key\\", \\"Orders\\".\\"timestamp\\" AS \\"timestamp\\", \\"Orders\\".\\"timestamptz\\" AS \\"timestamptz\\" FROM (SELECT \\"Orders\\".\\"orderkey\\" AS \\"orderkey\\", \\"Orders\\".\\"custkey\\" AS \\"custkey\\", \\"Orders\\".\\"orderstatus\\" AS \\"orderstatus\\", \\"Orders\\".\\"totalprice\\" AS \\"totalprice\\", \\"Orders\\".\\"orderdate\\" AS \\"orderdate\\", \\"Orders\\".\\"order_cust_key\\" AS \\"order_cust_key\\", \\"Orders\\".\\"timestamp\\" AS \\"timestamp\\", \\"Orders\\".\\"timestamptz\\" AS \\"timestamptz\\" FROM (SELECT o_orderkey AS \\"orderkey\\", o_custkey AS \\"custkey\\", o_orderstatus AS \\"orderstatus\\", o_totalprice AS \\"totalprice\\", o_orderdate AS \\"orderdate\\", CONCAT(o_orderkey, '_', o_custkey) AS \\"order_cust_key\\", CAST('2024-01-01T23:59:59' AS TIMESTAMP) AS \\"timestamp\\", CAST('2024-01-01T23:59:59' AS TIMESTAMPTZ) AS \\"timestamptz\\" FROM (SELECT * FROM public.orders) AS \\"Orders\\") AS \\"Orders\\") AS \\"Orders\\") SELECT orderkey, order_cust_key FROM \\"Orders\\" LIMIT 1"'''
-        )
