+++ ibis-server/tests/routers/v2/connector/test_mysql.py
@@ -3,7 +3,8 @@ import base64
 import orjson
 import pytest
 from fastapi.testclient import TestClient
-from testcontainers.mssql import SqlServerContainer
+from sqlalchemy import text
+from testcontainers.mysql import MySqlContainer
 
 from app.main import app
 
@@ -11,7 +12,7 @@ client = TestClient(app)
 
 
 @pytest.fixture(scope="class")
-def mssql(request) -> SqlServerContainer:
+def mysql(request) -> MySqlContainer:
     def file_path(path: str) -> str:
         import os
 
@@ -20,36 +21,43 @@ def mssql(request) -> SqlServerContainer:
     import pandas as pd
     import sqlalchemy
 
-    mssql = SqlServerContainer(
-        "mcr.microsoft.com/mssql/server:2019-latest", dialect="mssql+pyodbc"
-    )
-    mssql.start()
-    connection_url = mssql.get_connection_url()
-    engine = sqlalchemy.create_engine(f"{connection_url}?driver=FreeTDS")
-    pd.read_parquet(file_path("../../resource/tpch/data/orders.parquet")).to_sql(
+    mysql = MySqlContainer("mysql:8.0")
+    mysql.start()
+    connection_url = mysql.get_connection_url()
+    engine = sqlalchemy.create_engine(connection_url)
+    pd.read_parquet(file_path("../../../resource/tpch/data/orders.parquet")).to_sql(
         "orders", engine, index=False
     )
-    pd.read_parquet(file_path("../../resource/tpch/data/customer.parquet")).to_sql(
+    pd.read_parquet(file_path("../../../resource/tpch/data/customer.parquet")).to_sql(
         "customer", engine, index=False
     )
+    conn = engine.connect()
+    conn.execute(text("ALTER TABLE orders ADD PRIMARY KEY(o_orderkey);"))
+    conn.execute(text("ALTER TABLE customer Add PRIMARY KEY(c_custkey);"))
+    conn.execute(
+        text(
+            "ALTER TABLE orders ADD FOREIGN KEY (o_custkey) REFERENCES customer(c_custkey);"
+        )
+    )
+    conn.close()
 
     def stop_pg():
-        mssql.stop()
+        mysql.stop()
 
     request.addfinalizer(stop_pg)
 
-    return mssql
+    return mysql
 
 
-@pytest.mark.mssql
-class TestMSSql:
+@pytest.mark.mysql
+class TestMySql:
     manifest = {
         "catalog": "my_catalog",
         "schema": "my_schema",
         "models": [
             {
                 "name": "Orders",
-                "refSql": "select * from dbo.orders",
+                "refSql": "select * from orders",
                 "columns": [
                     {"name": "orderkey", "expression": "o_orderkey", "type": "integer"},
                     {"name": "custkey", "expression": "o_custkey", "type": "integer"},
@@ -84,7 +92,7 @@ class TestMSSql:
             },
             {
                 "name": "Customer",
-                "refSql": "select * from dbo.customer",
+                "refSql": "select * from customer",
                 "columns": [
                     {"name": "custkey", "expression": "c_custkey", "type": "integer"},
                     {"name": "name", "expression": "c_name", "type": "varchar"},
@@ -97,24 +105,24 @@ class TestMSSql:
     manifest_str = base64.b64encode(orjson.dumps(manifest)).decode("utf-8")
 
     @staticmethod
-    def to_connection_info(mssql: SqlServerContainer):
+    def to_connection_info(mysql: MySqlContainer):
         return {
-            "host": mssql.get_container_host_ip(),
-            "port": mssql.get_exposed_port(mssql.port),
-            "user": mssql.username,
-            "password": mssql.password,
-            "database": mssql.dbname,
+            "host": mysql.get_container_host_ip(),
+            "port": mysql.get_exposed_port(mysql.port),
+            "user": mysql.username,
+            "password": mysql.password,
+            "database": mysql.dbname,
         }
 
     @staticmethod
-    def to_connection_url(mssql: SqlServerContainer):
-        info = TestMSSql.to_connection_info(mssql)
-        return f"mssql://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
+    def to_connection_url(mysql: MySqlContainer):
+        info = TestMySql.to_connection_info(mysql)
+        return f"mysql://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
 
-    def test_query(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -143,16 +151,13 @@ class TestMSSql:
             "orderdate": "object",
             "order_cust_key": "object",
             "timestamp": "datetime64[ns]",
-            "timestamptz": "datetime64[ns, UTC]",
+            "timestamptz": "datetime64[ns]",
         }
 
-    @pytest.mark.skip(
-        reason="ibis does not support mssql using connection url, wait fix PR in ibis repo"
-    )
-    def test_query_with_connection_url(self, mssql: SqlServerContainer):
-        connection_url = self.to_connection_url(mssql)
+    def test_query_with_connection_url(self, mysql: MySqlContainer):
+        connection_url = self.to_connection_url(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={
                 "connectionInfo": {"connectionUrl": connection_url},
                 "manifestStr": self.manifest_str,
@@ -166,10 +171,10 @@ class TestMSSql:
         assert result["data"][0][0] == 1
         assert result["dtypes"] is not None
 
-    def test_query_with_column_dtypes(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query_with_column_dtypes(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -194,7 +199,7 @@ class TestMSSql:
             "1996-01-02 00:00:00.000000",
             "1_370",
             "2024-01-01 23:59:59.000000",
-            "2024-01-01 23:59:59.000000 UTC",
+            "2024-01-01 23:59:59.000000",
         ]
         assert result["dtypes"] == {
             "orderkey": "int32",
@@ -207,10 +212,10 @@ class TestMSSql:
             "timestamptz": "object",
         }
 
-    def test_query_without_manifest(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query_without_manifest(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={
                 "connectionInfo": connection_info,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -223,10 +228,10 @@ class TestMSSql:
         assert result["detail"][0]["loc"] == ["body", "manifestStr"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_without_sql(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query_without_sql(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -238,7 +243,7 @@ class TestMSSql:
 
     def test_query_without_connection_info(self):
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             json={
                 "manifestStr": self.manifest_str,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -251,10 +256,10 @@ class TestMSSql:
         assert result["detail"][0]["loc"] == ["body", "connectionInfo"]
         assert result["detail"][0]["msg"] == "Field required"
 
-    def test_query_with_dry_run(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query_with_dry_run(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -264,10 +269,10 @@ class TestMSSql:
         )
         assert response.status_code == 204
 
-    def test_query_with_dry_run_and_invalid_sql(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_query_with_dry_run_and_invalid_sql(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/query",
+            url="/v2/connector/mysql/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -278,10 +283,10 @@ class TestMSSql:
         assert response.status_code == 422
         assert response.text is not None
 
-    def test_validate_with_unknown_rule(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_validate_with_unknown_rule(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/validate/unknown_rule",
+            url="/v2/connector/mysql/validate/unknown_rule",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -294,10 +299,10 @@ class TestMSSql:
             == "The rule `unknown_rule` is not in the rules, rules: ['column_is_valid']"
         )
 
-    def test_validate_rule_column_is_valid(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_validate_rule_column_is_valid(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -307,11 +312,11 @@ class TestMSSql:
         assert response.status_code == 204
 
     def test_validate_rule_column_is_valid_with_invalid_parameters(
-        self, mssql: SqlServerContainer
+        self, mysql: MySqlContainer
     ):
-        connection_info = self.to_connection_info(mssql)
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -321,7 +326,7 @@ class TestMSSql:
         assert response.status_code == 422
 
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -331,11 +336,11 @@ class TestMSSql:
         assert response.status_code == 422
 
     def test_validate_rule_column_is_valid_without_parameters(
-        self, mssql: SqlServerContainer
+        self, mysql: MySqlContainer
     ):
-        connection_info = self.to_connection_info(mssql)
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -346,11 +351,11 @@ class TestMSSql:
         assert result["detail"][0]["msg"] == "Field required"
 
     def test_validate_rule_column_is_valid_without_one_parameter(
-        self, mssql: SqlServerContainer
+        self, mysql: MySqlContainer
     ):
-        connection_info = self.to_connection_info(mssql)
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -361,7 +366,7 @@ class TestMSSql:
         assert response.text == "Missing required parameter: `columnName`"
 
         response = client.post(
-            url="/v2/ibis/mssql/validate/column_is_valid",
+            url="/v2/connector/mysql/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -371,10 +376,10 @@ class TestMSSql:
         assert response.status_code == 422
         assert response.text == "Missing required parameter: `modelName`"
 
-    def test_metadata_list_tables(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_metadata_list_tables(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/metadata/tables",
+            url="/v2/connector/mysql/metadata/tables",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
@@ -386,10 +391,18 @@ class TestMSSql:
         assert result["description"] is not None
         assert result["properties"] is not None
 
-    def test_metadata_list_constraints(self, mssql: SqlServerContainer):
-        connection_info = self.to_connection_info(mssql)
+    def test_metadata_list_constraints(self, mysql: MySqlContainer):
+        connection_info = self.to_connection_info(mysql)
         response = client.post(
-            url="/v2/ibis/mssql/metadata/constraints",
+            url="/v2/connector/mysql/metadata/constraints",
             json={"connectionInfo": connection_info},
         )
         assert response.status_code == 200
+
+        result = response.json()[0]
+        assert result["constraintName"] is not None
+        assert result["constraintType"] is not None
+        assert result["constraintTable"] is not None
+        assert result["constraintColumn"] is not None
+        assert result["constraintedTable"] is not None
+        assert result["constraintedColumn"] is not None
