+++ ibis-server/tests/routers/v2/connector/test_snowflake.py
@@ -9,32 +9,33 @@ from app.main import app
 client = TestClient(app)
 
 
-@pytest.mark.bigquery
-class TestBigquery:
+@pytest.mark.snowflake
+class TestSnowflake:
     manifest = {
         "catalog": "my_catalog",
         "schema": "my_schema",
         "models": [
             {
                 "name": "Orders",
-                "refSql": "select * from tpch_tiny.orders",
+                "properties": {},
+                "refSql": "select * from TPCH_SF1.ORDERS",
                 "columns": [
-                    {"name": "orderkey", "expression": "o_orderkey", "type": "integer"},
-                    {"name": "custkey", "expression": "o_custkey", "type": "integer"},
+                    {"name": "orderkey", "expression": "O_ORDERKEY", "type": "integer"},
+                    {"name": "custkey", "expression": "O_CUSTKEY", "type": "integer"},
                     {
                         "name": "orderstatus",
-                        "expression": "o_orderstatus",
+                        "expression": "O_ORDERSTATUS",
                         "type": "varchar",
                     },
                     {
                         "name": "totalprice",
-                        "expression": "o_totalprice",
+                        "expression": "O_TOTALPRICE",
                         "type": "float",
                     },
-                    {"name": "orderdate", "expression": "o_orderdate", "type": "date"},
+                    {"name": "orderdate", "expression": "O_ORDERDATE", "type": "date"},
                     {
                         "name": "order_cust_key",
-                        "expression": "concat(o_orderkey, '_', o_custkey)",
+                        "expression": "concat(O_ORDERKEY, '_', O_CUSTKEY)",
                         "type": "varchar",
                     },
                     {
@@ -52,10 +53,10 @@ class TestBigquery:
             },
             {
                 "name": "Customer",
-                "refSql": "select * from tpch_tiny.customer",
+                "refSql": "select * from TPCH_SF1.CUSTOMER",
                 "columns": [
-                    {"name": "custkey", "expression": "c_custkey", "type": "integer"},
-                    {"name": "name", "expression": "c_name", "type": "varchar"},
+                    {"name": "custkey", "expression": "C_CUSTKEY", "type": "integer"},
+                    {"name": "name", "expression": "C_NAME", "type": "varchar"},
                 ],
                 "primaryKey": "custkey",
             },
@@ -69,19 +70,21 @@ class TestBigquery:
         import os
 
         return {
-            "project_id": os.getenv("TEST_BIG_QUERY_PROJECT_ID"),
-            "dataset_id": "tpch_tiny",
-            "credentials": os.getenv("TEST_BIG_QUERY_CREDENTIALS_BASE64_JSON"),
+            "user": os.getenv("SNOWFLAKE_USER"),
+            "password": os.getenv("SNOWFLAKE_PASSWORD"),
+            "account": os.getenv("SNOWFLAKE_ACCOUNT"),
+            "database": "SNOWFLAKE_SAMPLE_DATA",
+            "schema": "TPCH_SF1",
         }
 
     def test_query(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" ORDER BY orderkey LIMIT 1',
+                "sql": 'SELECT * FROM "Orders" ORDER BY "orderkey" LIMIT 1',
             },
         )
         assert response.status_code == 200
@@ -90,11 +93,11 @@ class TestBigquery:
         assert len(result["data"]) == 1
         assert result["data"][0] == [
             1,
-            370,
+            36901,
             "O",
-            172799.49,
+            173665.47,
             820540800000,
-            "1_370",
+            "1_36901",
             1704153599000,
             1704153599000,
         ]
@@ -102,7 +105,7 @@ class TestBigquery:
             "orderkey": "int64",
             "custkey": "int64",
             "orderstatus": "object",
-            "totalprice": "float64",
+            "totalprice": "object",
             "orderdate": "object",
             "order_cust_key": "object",
             "timestamp": "datetime64[ns]",
@@ -112,11 +115,11 @@ class TestBigquery:
     def test_query_with_column_dtypes(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
-                "sql": 'SELECT * FROM "Orders" ORDER BY orderkey LIMIT 1',
+                "sql": 'SELECT * FROM "Orders" ORDER BY "orderkey" LIMIT 1',
                 "columnDtypes": {
                     "totalprice": "float",
                     "orderdate": "datetime64",
@@ -131,11 +134,11 @@ class TestBigquery:
         assert len(result["data"]) == 1
         assert result["data"][0] == [
             1,
-            370,
+            36901,
             "O",
-            172799.49,
+            173665.47,
             "1996-01-02 00:00:00.000000",
-            "1_370",
+            "1_36901",
             "2024-01-01 23:59:59.000000",
             "2024-01-01 23:59:59.000000 UTC",
         ]
@@ -153,7 +156,7 @@ class TestBigquery:
     def test_query_without_manifest(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             json={
                 "connectionInfo": connection_info,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -169,7 +172,7 @@ class TestBigquery:
     def test_query_without_sql(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -181,7 +184,7 @@ class TestBigquery:
 
     def test_query_without_connection_info(self):
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             json={
                 "manifestStr": self.manifest_str,
                 "sql": 'SELECT * FROM "Orders" LIMIT 1',
@@ -197,7 +200,7 @@ class TestBigquery:
     def test_query_with_dry_run(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -210,7 +213,7 @@ class TestBigquery:
     def test_query_with_dry_run_and_invalid_sql(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/query",
+            url="/v2/connector/snowflake/query",
             params={"dryRun": True},
             json={
                 "connectionInfo": connection_info,
@@ -224,7 +227,7 @@ class TestBigquery:
     def test_validate_with_unknown_rule(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/validate/unknown_rule",
+            url="/v2/connector/snowflake/validate/unknown_rule",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -240,7 +243,7 @@ class TestBigquery:
     def test_validate_rule_column_is_valid(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -252,7 +255,7 @@ class TestBigquery:
     def test_validate_rule_column_is_valid_with_invalid_parameters(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -262,7 +265,7 @@ class TestBigquery:
         assert response.status_code == 422
 
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -274,7 +277,7 @@ class TestBigquery:
     def test_validate_rule_column_is_valid_without_parameters(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={"connectionInfo": connection_info, "manifestStr": self.manifest_str},
         )
         assert response.status_code == 422
@@ -287,7 +290,7 @@ class TestBigquery:
     def test_validate_rule_column_is_valid_without_one_parameter(self):
         connection_info = self.get_connection_info()
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -298,7 +301,7 @@ class TestBigquery:
         assert response.text == "Missing required parameter: `columnName`"
 
         response = client.post(
-            url="/v2/ibis/bigquery/validate/column_is_valid",
+            url="/v2/connector/snowflake/validate/column_is_valid",
             json={
                 "connectionInfo": connection_info,
                 "manifestStr": self.manifest_str,
@@ -307,19 +310,3 @@ class TestBigquery:
         )
         assert response.status_code == 422
         assert response.text == "Missing required parameter: `modelName`"
-
-    def test_metadata_list_tables(self):
-        connection_info = self.get_connection_info()
-        response = client.post(
-            url="/v2/ibis/bigquery/metadata/tables",
-            json={"connectionInfo": connection_info},
-        )
-        assert response.status_code == 200
-
-    def test_metadata_list_constraints(self):
-        connection_info = self.get_connection_info()
-        response = client.post(
-            url="/v2/ibis/bigquery/metadata/constraints",
-            json={"connectionInfo": connection_info},
-        )
-        assert response.status_code == 200
