+++ ibis-server/tests/routers/v2/test_analysis.py
